requirements.txt
# AI & ML
torch
torchvision
transformers
ollama
langchain
langchain-core
langchain-community
langchain-ollama
google-generativeai

# Data Science
numpy
pandas
matplotlib
seaborn

# Web & Scraping
fastapi
uvicorn[standard]
selenium
webdriver-manager
beautifulsoup4
duckduckgo-search
ddgs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Media & Processing
moviepy==1.0.3
Pillow
exifread  # pure-Python EXIF fallback; installing system `exiftool` is recommended for full feature parity
pillow-heif
opencv-python
imagehash
weasyprint
markdown2

# Utils
python-dotenv
pyyaml
python-dateutil
pytest
pytest-cov
requests
